+++ index.js
@@ -10,6 +10,7 @@ import cryptokey from './types/cryptokey.js';
 import dataview from './types/dataview.js';
 import date from './types/date.js';
 import domexception from './types/domexception.js';
+import dommatrix from './types/dommatrix.js';
 import dompoint from './types/dompoint.js';
 import domrect from './types/domrect.js';
 import error from './types/error.js';
@@ -50,10 +51,10 @@ import universal from './presets/universal.js';
 /* types */
 export {
     arraybuffer, bigintObject, bigint, blob, cloneable, cryptokey, dataview,
-    date, domexception, dompoint, domrect, error, errors, file, filelist,
-    imagebitmap, imagedata, infinity, intlTypes, map, nan, negativeInfinity,
-    negativeZero, nonbuiltinIgnore, primitiveObjects, regexp, resurrectable,
-    set, typedArraysSocketio, typedArrays, undef, userObject
+    date, domexception, dommatrix, dompoint, domrect, error, errors, file,
+    filelist, imagebitmap, imagedata, infinity, intlTypes, map, nan,
+    negativeInfinity, negativeZero, nonbuiltinIgnore, primitiveObjects, regexp,
+    resurrectable, set, typedArraysSocketio, typedArrays, undef, userObject
 };
 
 /* presets */