+++ test/helpers/test-environment.js
@@ -118,6 +118,62 @@ class DOMPoint {
 // @ts-expect-error Not an issue
 globalThis.DOMPoint = DOMPoint;
 
+// No constructor in JSDom
+// globalThis.DOMMatrix = window.DOMMatrix;
+/**
+ * DOMMatrix class.
+ */
+class DOMMatrix {
+    /* eslint-disable class-methods-use-this -- Not needed */
+    /**
+     * @returns {string}
+     */
+    get [Symbol.toStringTag] () {
+        /* eslint-enable class-methods-use-this -- Not needed */
+        return 'DOMMatrix';
+    }
+    /**
+     * @param {[number, number, number, number, number, number]|
+     *   [
+     *     number, number, number, number,
+     *     number, number, number, number,
+     *     number, number, number, number,
+     *     number, number, number, number
+     * ]} init
+     */
+    constructor (init) {
+        if (typeof init[6] !== 'number') {
+            this.is2D = true;
+            this.a = init[0];
+            this.b = init[1];
+            this.c = init[2];
+            this.d = init[3];
+            this.e = init[4];
+            this.f = init[5];
+            return;
+        }
+        this.is2D = false;
+        this.m11 = init[0];
+        this.m12 = init[1];
+        this.m13 = init[2];
+        this.m14 = init[3];
+        this.m21 = init[4];
+        this.m22 = init[5];
+        this.m23 = init[6];
+        this.m24 = init[7];
+        this.m31 = init[8];
+        this.m32 = init[9];
+        this.m33 = init[10];
+        this.m34 = init[11];
+        this.m41 = init[12];
+        this.m42 = init[13];
+        this.m43 = init[14];
+        this.m44 = init[15];
+    }
+}
+// @ts-expect-error Not an issue
+globalThis.DOMMatrix = DOMMatrix;
+
 globalThis.performance = window.performance;
 
 try {